Make sure slots to be used by VM are pre-determined
This commit introduced changes where we ensure that when tablet is added for VNC, libvirt does not automatically assign a slot but we do that by specifying a slot. This helps if we want to use some other slots via command line bhyve arguments, a use case is PCI passthrough devices wchich libvirt is not supporting yet for bhyve driver and we would be using command line args. If we don't have slots pre-defined, they could potentially clash with the ones we want to use with passthrough devies.